+++ hpc/nvidia/cuda/fakeCUDA/frontend/lexerExtension.cpp
@@ -1,0 +1,5 @@
+#include "syntax.hpp"
+#include "clang/Basic/TokenKinds.h"
+#include "clang/Lex/Preprocessor.h"
+
+using namespace clang;